feat: add config systems

+++ src/client/kotlin/org/infinite/libs/graphics/graphics2d/structs/StrokeStyle.kt
@@ -2,5 +2,5 @@ package org.infinite.libs.graphics.graphics2d.structs
 
 data class StrokeStyle(
     var color: Int,
-    var width: Double,
+    var width: Float,
 )
